Created a pattern using the core WP Query Block output, adding our own classes to modify
I copied the WP output that I built using the WP Core Query block and then created a pattern and added the classes for modifications.

I also included styling for the other default core blocks being used in the query. You'll notice in the pattern, we need to define the blocks we will use, one of which being the Column block so we added that back in (removed the removal line from editor.js)